git-svn-id: https://cleartoolkit.svn.sourceforge.net/svnroot/cleartoolkit/HTML5/trunk@1395 9db0787f-aee9-4a4d-a695-80f0cfb65743

+++ com.farata.cleardatabuilder.extjs/resources/mybatis_sample/src/com/farata/example/mapper/AssociateMapper.java
@@ -19,7 +19,7 @@ public interface AssociateMapper {
 		@Result(property="id", column="id"),
 		@Result(property="associateName", column="associate")
 	})
-	List<AssociateDTO> getAssociates(Long company_id);
+	List<AssociateDTO> getAssociates(Integer company_id);
 
 	@Insert("INSERT INTO company_associate (associate, company_id) VALUES (#{associateName}, #{companyId})")
 	//@SelectKey(statement = "call identity()", keyProperty = "id", before = false, resultType = Long.class)
